rectangle layout moved to Legend class; javadocs and removal of unused methods
