+++ jplotter/src/main/java/hageldave/jplotter/renderables/Legend.java
@@ -12,6 +12,8 @@ import java.util.LinkedHashMap;
 import java.util.LinkedList;
 import java.util.List;
 import java.util.Map;
+import java.util.function.BiConsumer;
+import java.util.function.Supplier;
 
 import org.w3c.dom.Document;
 import org.w3c.dom.Element;
@@ -25,7 +27,6 @@ import hageldave.jplotter.renderables.Points.PointDetails;
 import hageldave.jplotter.renderers.CompleteRenderer;
 import hageldave.jplotter.renderers.Renderer;
 import hageldave.jplotter.util.Annotations.GLContextRequired;
-import hageldave.jplotter.util.RectangleLayout;
 import hageldave.jplotter.util.Utils;
 
 /**
@@ -740,4 +741,92 @@ public class Legend implements Renderable, Renderer {
 		return false;
 	}
 
+	
+	/**
+	 * Class for arranging rectangles inspired by {@link java.awt.FlowLayout}.
+	 * Used to layout Legend elements.
+	 */
+	private static class RectangleLayout {
+
+		static abstract class LayoutComp {
+			int x,y;
+			abstract Dimension getSize();
+			abstract void apply();
+		}
+		
+		ArrayList<LayoutComp> comps = new ArrayList<>();
+		int width, height;
+		
+		Dimension getLayoutSize(){
+			return new Dimension(width, height);
+		}
+		
+		RectangleLayout addComponent(Supplier<Dimension> dim, BiConsumer<Integer, Integer> setLocation){
+			this.comps.add(new LayoutComp() {
+				@Override
+				Dimension getSize() {
+					return dim.get();
+				}
+				
+				@Override
+				void apply() {
+					setLocation.accept(x, y);
+				}
+			});
+			return this;
+		}
+		
+		RectangleLayout calculateFlowLayout(final int maxWidth, int hspace, int vspace){
+			int currX=0, currY=0, lineHeight=0;
+			this.width=this.height=0;
+			for(int i=0; i<comps.size(); i++){
+				LayoutComp c = comps.get(i);
+				Dimension size = c.getSize();
+				if(i==0){
+					c.x=currX;
+					c.y=currY;
+					currX+=size.width+hspace;
+					lineHeight = size.height;
+					this.width = currX-hspace;
+				} else {
+					// enough space
+					if(maxWidth-currX >= size.width){
+						c.x=currX;
+						c.y=currY;
+						currX+=size.width+hspace;
+						lineHeight = Math.max(lineHeight, size.height);
+						this.width = Math.max(this.width, currX-hspace);
+					// next line
+					} else {
+						currX=0;
+						currY+=lineHeight+vspace;
+						c.x=currX;
+						c.y=currY;
+						currX+=size.width+hspace;
+						lineHeight = size.height;
+						this.width = Math.max(this.width, currX-hspace);
+					}
+				}
+			}
+			this.height = currY+lineHeight;
+			return this;
+		}
+		
+		RectangleLayout flipYAxis(int height){
+			comps.forEach(c->c.y=height-c.y);
+			return this;
+		}
+		
+		RectangleLayout translate(int dx, int dy){
+			comps.forEach(c->{c.x+=dx;c.y+=dy;}); 
+			return this;
+		}
+		
+		RectangleLayout apply(){
+			comps.forEach(LayoutComp::apply);
+			return this;
+		}
+		
+	}
+	
 }
