+++ jplotter/src/main/java/hageldave/jplotter/util/BarycentricGradientPaint.java
@@ -415,5 +415,4 @@ public class BarycentricGradientPaint implements Paint {
 
 	}
 
-
 }
